+++ actionmailer/lib/action_mailer/preview.rb
@@ -25,31 +25,7 @@ module ActionMailer
       mattr_accessor :preview_interceptors, instance_writer: false, default: [ActionMailer::InlinePreviewInterceptor]
     end
 
-    def preview_path
-      ActionMailer.deprecator.warn(<<-MSG.squish)
-        Using preview_path option is deprecated and will be removed in Rails 7.2.
-        Please use preview_paths instead.
-      MSG
-      self.class.preview_paths.first
-    end
-
     module ClassMethods
-      def preview_path=(value)
-        ActionMailer.deprecator.warn(<<-MSG.squish)
-          Using preview_path= option is deprecated and will be removed in Rails 7.2.
-          Please use preview_paths= instead.
-        MSG
-        self.preview_paths << value
-      end
-
-      def preview_path
-        ActionMailer.deprecator.warn(<<-MSG.squish)
-          Using preview_path option is deprecated and will be removed in Rails 7.2.
-          Please use preview_paths instead.
-        MSG
-        self.preview_paths.first
-      end
-
       # Register one or more Interceptors which will be called before mail is previewed.
       def register_preview_interceptors(*interceptors)
         interceptors.flatten.compact.each { |interceptor| register_preview_interceptor(interceptor) }