+++ actionmailer/CHANGELOG.md
@@ -1,2 +1,5 @@
+*   Remove deprecated option `config.action_mailer.preview_path`.
+
+    *Mike Dalessio*
 
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/actionmailer/CHANGELOG.md) for previous changes.